requirements.txt
google-adk>=0.2.0
google-cloud-storage>=2.10.0
gspread>=5.12.0
oauth2client>=4.1.3
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0

